Increased ticket limit to uint16_t

+++ tls/s2n_connection.h
@@ -332,7 +332,7 @@ struct s2n_connection {
     bool send_in_use;
     bool recv_in_use;
     
-    uint8_t tickets_to_send;
+    uint16_t tickets_to_send;
 };
 
 int s2n_connection_is_managed_corked(const struct s2n_connection *s2n_connection);
